make primary & foreign keys long instead of int data type

+++ RestaurantReservation.Db/Models/Employee.cs
@@ -2,8 +2,8 @@
 
 public class Employee
 {
-    public int EmployeeId { get; set; }
-    public int RestaurantId { get; set; }
+    public long EmployeeId { get; set; }
+    public long RestaurantId { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
     public EmployeePosition Position { get; set; }
