+++ RestaurantReservation.Db/Models/Order.cs
@@ -2,9 +2,9 @@
 
 public class Order
 {
-    public int OrderId { get; set; }
-    public int ReservationId { get; set; }
-    public int? EmployeeId { get; set; }
+    public long OrderId { get; set; }
+    public long ReservationId { get; set; }
+    public long? EmployeeId { get; set; }
     public DateTime OrderDate { get; set; }
     public int TotalAmount { get; set; }
